requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anthropic==0.2.10
anyio==3.7.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.4.1
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.3
async-timeout==4.0.2
attrs==23.1.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.6
azure-cosmos==4.5.1
azure-functions==1.17.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.0
bentoml==1.0.22
bleach==6.0.0
blinker==1.6.2
build==0.10.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.1
chromadb==0.4.2
circus==0.18.0
clarifai==9.1.0
clarifai-grpc==9.5.0
click==8.1.3
click-option-group==0.5.6
cloudpickle==2.2.1
cohere==3.10.0
coloredlogs==15.0.1
comm==0.1.3
contextlib2==21.6.0
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
dataclasses-json==0.5.8
datasets==2.13.1
debugpy==1.6.7
decorator==5.1.1
deepmerge==1.1.0
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.6
distro==1.8.0
dm-tree==0.1.8
et-xmlfile==1.1.0
etils==1.5.2
executing==1.2.0
fastapi==0.96.0
fastjsonschema==2.17.1
ffmpy==0.3.0
filelock==3.12.0
filetype==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.39.4
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.5.0
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.32
google==3.0.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
gpt-index==0.4.24
gradio==3.34.0
gradio_client==0.2.6
grpcio==1.56.0
grpcio-health-checking==1.48.2
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
homebrew==0.2.5
html5lib==1.1
httpcore==0.17.2
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.1
importlib-resources==6.0.0
inflection==0.5.1
ipykernel==6.24.0
ipython==8.14.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
keras==2.15.0
kiwisolver==1.4.4
langchain==0.0.305
langchainplus-sdk==0.0.17
langsmith==0.0.41
libclang==16.0.6
linkify-it-py==2.0.2
llama-index==0.5.6
lxml==4.9.3
manifest-ml==0.0.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.1
ml-dtypes==0.2.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.14
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
nlpcloud==1.0.42
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.15.1
openai==1.3.7
openapi-schema-pydantic==1.2.4
openllm==0.1.13
openlm==0.0.5
openpyxl==3.1.2
opentelemetry-api==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-aiohttp-client==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-instrumentation-grpc==0.38b0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
opt-einsum==3.3.0
optimum==1.8.8
orjson==3.9.0
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip-requirements-parser==32.0.1
pip-tools==6.13.0
platformdirs==3.8.1
posthog==3.0.1
praw==7.7.0
prawcore==2.3.0
prometheus-client==0.17.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.2.0
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.9
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.3.0
Pygments==2.15.1
Pympler==1.0.1
pynvml==11.5.0
pypandoc==1.11
pyparsing==3.0.9
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytube==12.1.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.1.0
redis==4.5.5
regex==2023.6.3
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
schema==0.7.5
scikit-learn==1.3.2
scipy==1.11.4
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.1.99
simple-di==0.1.5
six==1.16.0
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.48
sqlitedict==2.1.0
stack-data==0.6.2
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3.1
uc-micro-py==1.0.2
unstructured==0.8.1
update-checker==0.18.0
uri-template==1.3.0
urllib3==1.26.15
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
Wand==0.6.11
watchfiles==0.19.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
xxhash==3.2.0
yarl==1.8.2
yfinance==0.2.22
youtube-dl==2021.12.17
zipp==3.15.0
